Refactor AC & PPO
