+++ examples/rl/cim/policy_trainer.py
@@ -11,6 +11,10 @@ if algorithm == "ac":
     from .algorithms.ac import get_ac, get_policy
     policy_creator = {f"ac_{i}.policy": partial(get_policy, state_dim, action_num) for i in range(4)}
     trainer_creator = {f"ac_{i}": partial(get_ac, state_dim) for i in range(4)}
+elif algorithm == "ppo":
+    from .algorithms.ac import get_ppo, get_policy
+    policy_creator = {f"ppo_{i}.policy": partial(get_policy, state_dim, action_num) for i in range(4)}
+    trainer_creator = {f"ppo_{i}": partial(get_ppo, state_dim) for i in range(4)}
 elif algorithm == "dqn":
     from .algorithms.dqn import get_dqn, get_policy
     policy_creator = {f"dqn_{i}.policy": partial(get_policy, state_dim, action_num) for i in range(4)}
