+++ maro/rl/training/algorithms/__init__.py
@@ -5,10 +5,12 @@ from .ac import DiscreteActorCriticParams, DiscreteActorCriticTrainer
 from .ddpg import DDPGParams, DDPGTrainer
 from .dqn import DQNParams, DQNTrainer
 from .maddpg import DiscreteMADDPGParams, DiscreteMADDPGTrainer
+from .ppo import DiscretePPOParams, DiscretePPOTrainer
 
 __all__ = [
     "DiscreteActorCriticTrainer", "DiscreteActorCriticParams",
     "DDPGTrainer", "DDPGParams",
     "DQNTrainer", "DQNParams",
     "DiscreteMADDPGTrainer", "DiscreteMADDPGParams",
+    "DiscretePPOParams", "DiscretePPOTrainer",
 ]
